// src/app/Interfaces/what.ts
export interface whatContentTypes {
    slugName: string;
    mainHeadline: string;
    description: string;
    serviceHeading: string;
    serviceTab1Heading: string;
    serviceTab1Desc: string;
    serviceTab2Heading: string;
    serviceTab2Desc: string;
    serviceTab3Heading: string;
    serviceTab3Desc: string;
    workHeading: string;
    workDesc: string;
    workStep1Heading: string;
    workStep1Desc: string;
    workStep2Heading: string;
    workStep2Desc: string;
    workStep3Heading: string;
    workStep3Desc: string;
    workStep4Heading: string;
    workStep4Desc: string;
  }
  